Add UI interaction with lasers
Created an example Canvas with example UI in the game. Made the Canvas Render Mode in World Space. You have to add a "Tracked Device Graphic Raycaster" component to the Canvas so that laser can interact with it. You must also replace the module script in event system with "XRUI Input Module"